assets/img/Resonique/requirements.txt
streamlit
torch
sentence-transformers
transformers
numpy
pillow
pinecone
google-generativeai
supabase 
pandas
requests
librosa
ipywidgets
wav2clip
matplotlib
json
functools